Reorganized and cleaned class RinchiOptions. Enums ReactionFileFormat and RinchiKeyType moved to separate java files
