+++ jna-rinchi-api/src/main/java/io/github/dan2097/jnarinchi/ReactionFileFormat.java
@@ -1,0 +1,7 @@
+package io.github.dan2097.jnarinchi;
+
+public enum ReactionFileFormat {	
+	RXN, 
+	RD, 
+	AUTO;
+}
