+++ jna-rinchi-api/src/main/java/io/github/dan2097/jnarinchi/RinchiKeyType.java
@@ -1,0 +1,18 @@
+package io.github.dan2097.jnarinchi;
+
+public enum RinchiKeyType 
+{
+	LONG ("L"), 
+	SHORT ("S"), 
+	WEB ("W");
+	
+	private final String shortDeignation;
+	
+	private RinchiKeyType (String shortDeignation) {
+		this.shortDeignation = shortDeignation; 
+	}
+	
+	public String getShortDeignation() {
+		return shortDeignation;
+	}
+}
